pkg/distro/distroutil/detect/detect_test.go
package detect

import (
	"strings"
	"testing"

	"gotest.tools/v3/assert"
)

func TestDistroID(t *testing.T) {
	assertFn := func(name string, etcOSRelease string) {
		r := strings.NewReader(etcOSRelease)
		id, err := distroID(r)
		assert.NilError(t, err)
		assert.Equal(t, name, id)
	}

	assertFn("debian", `PRETTY_NAME="Debian GNU/Linux 11 (bullseye)"
NAME="Debian GNU/Linux"
VERSION_ID="11"
VERSION="11 (bullseye)"
VERSION_CODENAME=bullseye
ID=debian
HOME_URL="https://www.debian.org/"
SUPPORT_URL="https://www.debian.org/support"
BUG_REPORT_URL="https://bugs.debian.org/"
`)
	assertFn("ubuntu", `PRETTY_NAME="Ubuntu 22.04.1 LTS"
NAME="Ubuntu"
VERSION_ID="22.04"
VERSION="22.04.1 LTS (Jammy Jellyfish)"
VERSION_CODENAME=jammy
ID=ubuntu
ID_LIKE=debian
HOME_URL="https://www.ubuntu.com/"
SUPPORT_URL="https://help.ubuntu.com/"
BUG_REPORT_URL="https://bugs.launchpad.net/ubuntu/"
PRIVACY_POLICY_URL="https://www.ubuntu.com/legal/terms-and-policies/privacy-policy"
UBUNTU_CODENAME=jammy
`)
}
